Added and tested route to modify dataset table

+++ megatrack/views.py
@@ -11,7 +11,7 @@ import megatrack.cache_utils as cu
 import megatrack.data_utils as du
 import megatrack.database_utils as dbu
 import time
-from megatrack import bcrypt
+from megatrack import bcrypt, db
 
 megatrack = Blueprint('megatrack', __name__)
 
@@ -69,22 +69,33 @@ def modify_tracts():
 
 @megatrack.route('/datasets', methods=['GET','POST','PUT','DELETE'])
 def modify_datasets():
+    # get auth header and split to get the token string
     auth_header = request.headers.get('Authorization')
-    if auth_header:
-        auth_token = auth_header.split(" ")[1]
-        if auth_token:
-            user_id = User.decode_auth_token(auth_token)
-            
+    auth_token = auth_header.split(" ")[1] if auth_header else ''
+    
+    if auth_token:
+        # if auth token sent, deocde to get user id
+        user_id = User.decode_auth_token(auth_token)
+        # get user from database
+        # is it necessary to check the user id exists in the database?
+        # or is it sufficient that the token could be decoded using our secret key?
+        # its useful to have the user name for logging purposes anyway 
+        user = User.query.filter(User.user_id == user_id).first()
+        
+        if user:
             if request.method == 'GET':
                 try:
                     datasets = Dataset.query.all()
                     response_object = {
-                        "datasets": datasets
+                        'message': 'Successfully retrieved dataset records',
+                        'datasets': datasets
                     }
-                    return make_response(response_object), 200
+                    current_app.logger.info(f'User {user.user_name} retrieved all dataset records.')
+                    return make_response(jsonify(response_object)), 200
                 except Exception as e:
+                    current_app.logger.error(f'Error occurred while user "{user.user_name}" was attempting to get all dataset records.')
                     current_app.logger.error(e)
-                    return 'An error occured while getting datasets.', 500
+                    return 'An error occurred while getting datasets.', 500
                 
             elif request.method == 'POST':
                 form = request.form
@@ -92,21 +103,30 @@ def modify_datasets():
                     dataset = Dataset(form['code'], form['name'], form['filePath'], form['queryParams'])
                     db.session.add(dataset)
                     db.session.commit()
-                    return 'New dataset successfully created.', 200
+                    current_app.logger.info(f'User "{user.user_name}" inserted a new dataset with code "{dataset.code}"')
+                    return 'New dataset successfully created.', 201
                 except Exception as e:
+                    db.session.rollback()
+                    current_app.logger.error(f'Error occurred while user "{user.user_name}" was attempting to insert dataset.')
                     current_app.logger.error(e)
-                    return 'An error occurred while created a dataset record.', 500
+                    return 'An error occurred while creating a dataset record.', 500
                 
             elif request.method == 'PUT':
                 form = request.form
                 try:
                     dataset = Dataset.query.filter(Dataset.code == form['code']).first()
-                    dataset.name = form['name']
-                    dataset.file_path = form['filePath']
-                    dataset.query_params = form['queryParams']
-                    db.session.commit()
-                    return 'Dataset successfully updated.', 200
+                    if dataset:
+                        dataset.name = form['name']
+                        dataset.file_path = form['filePath']
+                        dataset.query_params = form['queryParams']
+                        db.session.commit()
+                        current_app.logger.info(f'User {user.user_name} updated dataset {dataset.code}.')
+                        return 'Dataset successfully updated.', 200
+                    else:
+                        raise Exception('Can\'t update dataset that doesn\'t exist.')
                 except Exception as e:
+                    db.session.rollback()
+                    current_app.logger.error(f'Error occurred while user {user.user_name} was attempting to update dataset {dataset}.')
                     current_app.logger.error(e)
                     return 'An error occurred while updating dataset.', 500
             
@@ -114,15 +134,29 @@ def modify_datasets():
                 code = request.args['code']
                 if code:
                     try:
-                        Dataset.query.filter(Dataset.code == code).delete()
-                        db.session.commit()
+                        dataset = Dataset.query.filter(Dataset.code == code).first()
+                        if dataset:
+                            Dataset.query.filter(Dataset.code == code).delete()
+                            db.session.commit()
+                        else:
+                            raise Exception(f'Cannot delete dataset with code "{code}" since a record doesn\'t exist.')
+                        current_app.logger.info(f'User {user.user_name} deleted dataset {code}.')
                         return 'Dataset successfully deleted.', 200
                     except Exception as e:
+                        db.session.rollback()
+                        current_app.logger.error(f'Error occurred while user {user.user_name} was attempting to delete dataset with code {code}.')
                         current_app.logger.error(e)
                         return 'An error occurred while deleting dataset.', 500
                 else:
-                    current_app.logger.error('No code sent with dataset DELETE request')
-                    return 'No dataset code was sent with DELETE request.', 404
+                    current_app.logger.warn('No code sent with dataset DELETE request')
+                    return 'No dataset code was sent with DELETE request.', 400
+         
+        else:
+            current_app.logger.warn(f'No user with id "{user_id}" found in database.')
+            return 'Invalid user id passed with authentication token', 401
+    else:
+        current_app.logger.warn('No authentication token sent with request to /datasets')
+        return 'No authentication token sent with request.', 401
 
 @megatrack.route('/get_template')
 def get_template():
